refactor(my-copilot): group video system files by domain
Move video components into src/components/video with dedicated hooks and
stories subfolders, and update all imports/references across pages, components,
hooks, tests, and stories.
