+++ apps/my-copilot/src/components/video/hooks/use-copy-to-clipboard.ts
