+++ apps/my-copilot/src/components/video/hooks/use-detail-page-controller.ts
