+++ apps/my-copilot/src/components/video/hooks/use-shorts-feed-telemetry-adapter.ts
