+++ apps/my-copilot/src/components/video/stories/shorts-feed.stories.tsx
@@ -1,31 +1,32 @@
 import type { Meta, StoryObj } from "@storybook/nextjs";
 import { Box } from "@navikt/ds-react";
 import { relatedVideos } from "./storybook-video-fixtures";
-import { RelatedVideos } from "./related-videos";
+import { ShortsFeed } from "../shorts-feed";
 
 const meta = {
-  title: "Video/Panels/Related Videos",
-  component: RelatedVideos,
+  title: "Video/Pages/Home Shorts Feed",
+  component: ShortsFeed,
   tags: ["autodocs"],
   parameters: {
     docs: {
       description: {
         component:
-          "Panel med relaterte videoer i 9:16-format. Historiene bruker faktiske media-URL-er fra dev-bucket for realistisk layout- og innholdsvalidering.",
+          "Dette er hjemmesidens faktiske videospiller-feed. Samme player-surface brukes også på detaljsiden gjennom delt komponent.",
       },
     },
   },
   decorators: [
     (Story) => (
-      <Box padding="space-16" className="rounded-xl bg-black">
+      <Box className="max-w-[720px] rounded-xl bg-black" padding="space-16">
         <Story />
       </Box>
     ),
   ],
   args: {
     videos: relatedVideos,
+    initialVideoId: relatedVideos[1]?.id,
   },
-} satisfies Meta<typeof RelatedVideos>;
+} satisfies Meta<typeof ShortsFeed>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
