+++ apps/my-copilot/src/components/video/stories/storybook-video-fixtures.ts
