+++ apps/my-copilot/src/components/video/vertical-video-container.tsx
