+++ apps/my-copilot/src/components/video/video-accent.ts
